+++ tests/test_git_webhook.py
@@ -3,7 +3,7 @@ from django.core.urlresolvers import reverse
 from .factories import PageFactory
 from waliki.settings import WALIKI_DATA_DIR
 from waliki.models import Page
-from waliki.git import Git
+from waliki.git.models import Git
 import json
 import os
 import shutil
