+++ waliki/templates/base.html
@@ -1,5 +1,4 @@
 {% load staticfiles %}
-{% load url from future %}
 {% load waliki_tags %}
 <!DOCTYPE html>
 <html lang="{{ LANGUAGE_CODE }}">
