+++ waliki_project/waliki_project/templates/allauth/account/base.html
@@ -1,6 +1,4 @@
 {% extends "base.html" %}
-
-{% load url from future %}
 {% load i18n %}
 
 {% block content %}
