+++ waliki_project/waliki_project/templates/allauth/account/email.html
@@ -1,6 +1,5 @@
 {% extends "account/base.html" %}
 
-{% load url from future %}
 {% load i18n %}
 {% load bootstrap %}
