+++ waliki_project/waliki_project/templates/allauth/account/logout.html
@@ -1,6 +1,5 @@
 {% extends "account/base.html" %}
 
-{% load url from future %}
 {% load i18n %}
 
 {% block head_title %}{% trans "Sign Out" %}{% endblock %}
