+++ waliki_project/waliki_project/templates/allauth/account/signup.html
@@ -1,6 +1,5 @@
 {% extends "account/base.html" %}
 
-{% load url from future %}
 {% load bootstrap %}
 {% load i18n %}
